Dynamic joint trajectory plotting
This commit adds a script to plot the trajectories (positions and
velocities) executed during the change_trajectory method in
demo_interface.py. The change_trajectory method implements dynamic
trajectory planning and execution in which the robot plans one path
while executing another, and then executes the second plan mid-execution
of the first plan.

This commit also changes the planning_context.launch used in the
demo.launch file so that the urdf is more precise and plans can be
executed without falure.