Cpp demo for testing jitter with joint trajectory client
The JointTrajectoryClient can take interrupting joint position goals and
navigate to them without jittering. Thus, if I need to stop while
executing a trajectory I can send the JointTrajectoryClient a joint
position goal made up of my current joint positions to come to a
controlled stop at a specified joint position.

TODO: Currently this enacts some back-tracking because there is some
time delay in between when I query for my current position and when the
JointTrajectoryClient executes that "current" position. Another option
(2) would be to use a velocity controller that sends my velocities to
zero and doesn't worry about the positions. Or (3) I could look at the
plan I'm enacting and send a joint goal a few steps ahead. It is going
to be advantageous to specify what joint position I go to since that is
what my tree is built on so option 3 will likely be best.

Future Work: with RT-RRT*, I can send "stopping" commands to the
JointTrajectoryClient to go to nodes near my current joint position when
needed. This will allow me to sync up with the RT-RRT* tree and utilize
stored paths for planning the next trajectory.