Add cpp collision check testing code
Add cpp script for checking collisions with self and environment as well
as constraint violations.

Also add publish_obstacle.py to publish obstacles to the environment as
desired to make sure the collision checker is able to recognize
collisions with objects in the environment

TODO:
* Learn about bullet continuous collision checker and checking
collisions between 2 valid states.
* Take into account realistic joint constraints as I do for the real
robot demo
* Make the demo work with the /joint_state_publisher node rather than
fighting against it. (currently to run this script and visualize the
states as desired, the /joint_state_publisher node must be shut down)