Adds various demo and calibration changes
Change Panda joint limits in staged_panda.urdf to the joint limits
specified by the manufacturers
(https://frankaemika.github.io/docs/control_parameters.html). Ideally,
this should lead to the demo not running into joint limits so frequently
anymore.

Shorten planning time from 5 seconds to 2 seconds.

Publish box object of realistic size and offset the location in the
planning scene to account for the fact that the April Tag detection
returns a point on the surface of the box, not in the middle of the box.

Change orientation angle to 0 degrees so panda end effector is verticle
as it tracks the object through the workspace. This allows for easier
joint configurations during object tracking.

Comment out roslib.load_manifest in display_marker_publisher.py. This
line was causing errors when attempting to run the demo and appears
unnecessary.

Edit grasp_test_stationary.py to grip the box used in the demo.

Alter the sample_pose.py script for collecting calibration data. Once
run, it queries the depth camera for the target april tag location -
which should be on the box between the pandas grippers - and stores that
with the location of "panda_link_9" which should be at the desired
location of the target april tag in the URDF. This gives us sim vs real
data that can help us calibrate, the "panda_link_9" location being the
ground truth location of the box in the world frame, and the depth
camera output being the detected location that needs to be calibrated.

Add panda_link_9 in the staged_panda.urdf.