Implement dynamic/monitored RRT* demo
This commit contains various changes to make the RRT* demo more dynamic.

Previously, the robot would plan a path to a goal and have to wait until
the path was fully executed before checking for a new goal and planning
a new path.

These changes enable the robot to smoothly stop mid-trajectory, and plan
to a new/changed goal state.

Includes a simple filter method to make sure the goal state changed
significantly before stopping and replanning. This is necessary to
compensate for noise in the sensor data reading the goal states.
Note: may want to improve this filtration method in the future:
1. Sometimes the camera sends corrupted values that are very far from
   accurate, may want to include a buffer to make sure none of these get
   sent through as goals.
2. Sometimes the last value that makes it through the filter is while
   the box is in-motion. May want to allow a more precise goal to make
   it through the buffer after we detect the box has settled at a
   location.

Add formatting changes/removal of obsolete code from demo_interface

Add `simulation` launch argument and ROS parameter so that the
demo_interface can decide which method of smooth-stopping to employ:
* Hardware smooth-stop method doesn't work in simulation since the
  hardware controller node doesn't get launched.
* Simulation stopping method causes motor shuddering on hardware

Remove calibration files from git and add calibration directory to
.gitignore.

THES-76: Implement smooth-stopping method
THES-25: Add in ability to use different controller for sim/hardware
THES-29: Add in noise tolerance