Update and rename RRT_dynamic_interface
Rename RRT_dynamic_interface to RRT_star_monitor as it is not
technically an interface or dynamic. The class implemented in the file,
now called RRTstarMonitor, monitors a dynamic environment to help manage
the static planner.

Use smooth_stop method from the demo_interface in RRTstarMonitor so the
monitor can work on hardware or in simulation