Add Parellelized RTRRTstar controller
THES-82: Create external controller for Parallelized RTRRT*

This commit adds code to create the PRTRRTstar controller. There are 2
versions, one for simulation and one for hardware. Their behavior is the
same, the only difference is the method in which they communicate with
the controller client.

This controller waits for a current path to be published by the planner.
Once it recieves a current path it waits for the collision checker to
verify that the path is clear, then it will begin executing to the next
state in the path. Upon execution to the next state, it will notify both
the collision checker and the planner that it is moving to the next
state.

The collision checker reads this and begins checking the next step in
the current path for collisions. (TODO) The planner will read this and
move the current node to the next step in the current path and begin
expanding/rewiring from there.